MOM/mom_online/requirements.txt
flask
google-generativeai
markdown
xhtml2pdf
gunicorn
